Tracking API module service
Issue #2980550
