+++ drush.services.yml
@@ -1,6 +1,6 @@
 services:
   commerce_canadapost.commands:
     class: \Drupal\commerce_canadapost\Commands\Commands
-    arguments: ['@config.factory', '@logger.factory']
+    arguments: ['@commerce_canadapost.tracking_api']
     tags:
       - { name: drush.command }
